+++ share/dashboard_react/src/Pages/Dashboard/components/Logs/index.jsx
@@ -1,27 +1,10 @@
 import React, { useState } from 'react'
-import TagPill from '../../../components/TagPill'
-import { Box, Code, useColorMode } from '@chakra-ui/react'
+import TagPill from '../../../../components/TagPill'
+import { Box, Code } from '@chakra-ui/react'
+import styles from './styles.module.scss'
 
 function Logs({ logs }) {
-  const { colorMode } = useColorMode()
   const [isScrollable, setIsScrollable] = useState(true)
-  const styles = {
-    table: {
-      width: '100%',
-      borderCollapse: 'collapse'
-    },
-    tr: {
-      borderBottom: '1px solid',
-      borderColor: colorMode === 'light' ? '#E9E9E9' : ''
-    },
-    td: {
-      paddingTop: '3px',
-      paddingBottom: '3px'
-    },
-    timestamp: {
-      minWidth: '200px'
-    }
-  }
 
   const handleClick = () => {
     //setIsScrollable(true)
@@ -29,21 +12,21 @@ function Logs({ logs }) {
 
   return (
     <Box onClick={handleClick} w='100%' maxH='500px' overflow={isScrollable ? 'auto' : 'hidden'}>
-      <table style={styles.table}>
+      <table className={styles.table}>
         {logs
           ?.filter((log) => log.timestamp)
           .map((log, index) => {
             const levelColor =
               log.level === 'INFO' ? 'blue' : log.level === 'WARN' ? 'orange' : log.level === 'ERROR' ? 'red' : 'gray'
             return (
-              <tr key={index} style={styles.tr}>
-                <td style={{ ...styles.td, ...styles.timestamp }}>
+              <tr key={index} className={styles.tr}>
+                <td className={`${styles.td} ${styles.timestamp}`}>
                   <Code bg='transparent'>{log.timestamp}</Code>{' '}
                 </td>
-                <td style={styles.td}>
+                <td className={styles.td}>
                   <TagPill text={log.level} colorScheme={levelColor} />
                 </td>
-                <td style={{ ...styles.td, ...styles.text }}>
+                <td className={`${styles.td} ${styles.text}`}>
                   <Code bg='transparent'>{log.text}</Code>
                 </td>
               </tr>
