--- share/dashboard_react/src/Pages/Dashboard/components/Proxies.jsx/ProxyGrid/styles.module.scss
@@ -1,9 +1,0 @@
-.accordionHeader {
-  border-radius: 0;
-  padding: rem(6);
-  font-size: rem(15);
-}
-.accordionPanel {
-  border-radius: 0;
-  border: none;
-}
