+++ share/dashboard_react/src/Pages/Dashboard/components/Proxies/ProxyGrid/styles.module.scss
@@ -1,0 +1,30 @@
+.accordionHeader {
+  border-radius: 0;
+  padding: rem(6);
+  font-size: rem(15);
+}
+.accordionPanel {
+  border-radius: 0;
+  border: none;
+}
+
+.card {
+  border-radius: $border-radius-sm;
+  border: rem(1) solid;
+  border-color: var(--secondary-color);
+  width: 100%;
+  gap: 0;
+}
+
+.header {
+  text-align: center;
+  padding: rem(4);
+  background-color: var(--secondary-color);
+  border-top-left-radius: $border-radius-sm;
+  border-top-right-radius: $border-radius-sm;
+  color: var(--white-color) !important;
+  font-weight: bold;
+  p {
+    color: var(--white-color) !important;
+  }
+}
