+++ share/dashboard_react/src/Pages/Dashboard/components/Proxies/ProxyLogo/styles.module.scss
@@ -1,0 +1,6 @@
+.image {
+  width: rem(32);
+  height: rem(32);
+  margin: auto;
+  object-fit: cover;
+}
