+++ share/dashboard_react/src/Pages/Dashboard/components/Proxies/ProxyMenu.jsx
