+++ share/dashboard_react/src/Pages/Dashboard/components/Proxies/ProxyTable/styles.module.scss
@@ -1,0 +1,3 @@
+.tableContainer {
+  width: 100%;
+}
