+++ share/dashboard_react/src/Pages/Dashboard/components/Proxies/index.jsx
