+++ share/dashboard_react/src/Pages/Dashboard/index.jsx
@@ -5,30 +5,14 @@ import { useSelector } from 'react-redux'
 import ClusterWorkload from './components/ClusterWorkload'
 import { Flex } from '@chakra-ui/react'
 import AccordionComponent from '../../components/AccordionComponent/index.jsx'
-import ClusterTests from './components/ClusterTests'
 import Logs from './components/Logs'
 import DBServers from './components/DBServers'
-import { css } from '@emotion/react'
-import Proxies from './components/Proxies.jsx'
+import Proxies from './components/Proxies'
 
 function Dashboard({ selectedCluster, user }) {
   const {
     common: { isDesktop }
   } = useSelector((state) => state)
-
-  const styles = {
-    workloadPanel: {
-      position: 'relative',
-      minHeight: '125px',
-      top: '-25px'
-    },
-    workloadAccordion: {
-      // '& .chakra-collapse': {
-      //   height: '100px !important',
-      //   overflow: 'visible'
-      // }
-    }
-  }
 
   return (
     <Flex direction='column' gap='8px'>
@@ -44,8 +28,6 @@ function Dashboard({ selectedCluster, user }) {
       <AccordionComponent
         heading={'Cluster Workload'}
         body={<ClusterWorkload workload={selectedCluster?.workLoad} />}
-        sx={styles.workloadAccordion}
-        // panelSX={styles.workloadPanel}
       />
       {selectedCluster && (
         <AccordionComponent
@@ -58,7 +40,7 @@ function Dashboard({ selectedCluster, user }) {
       <AccordionComponent
         heading={'Proxies'}
         panelSX={{ overflowX: 'auto', p: 0 }}
-        body={<Proxies selectedCluster={selectedCluster} />}
+        body={<Proxies selectedCluster={selectedCluster} user={user} />}
       />
 
       <AccordionComponent heading={'Cluster Logs'} body={<Logs logs={selectedCluster?.log?.buffer} />} />
