+++ share/dashboard_react/src/components/Icons/CustomIcon/index.jsx
@@ -1,0 +1,9 @@
+import { Icon } from '@chakra-ui/react'
+import React from 'react'
+import styles from './styles.module.scss'
+
+function CustomIcon({ icon, color, fontSize = '1.5rem', fill }) {
+  return <Icon fontSize={fontSize} className={styles[color]} as={icon} fill={fill} />
+}
+
+export default CustomIcon
