+++ share/dashboard_react/src/components/Icons/CustomIcon/styles.module.scss
@@ -1,0 +1,10 @@
+.red {
+  fill: var(--danger-primary-color);
+}
+.green {
+  fill: var(--success-primary-color);
+}
+
+.orange {
+  fill: var(--warning-primary-color);
+}
